Reorganized bash_profile
Because this bash_profile is the one I use on multiple machines, there
are many places where I check which machine I'm on. In this revision, I
reogranized the code so that it now is easier to tell which code is
being executed on a specific machine, and which is being executed on all
machines.