rcm: Create links for the Andrew Unix machines
Over the summer, the result of `hostname` on the Andrew Unix machines
changed, meaning that if we do rcup/rcdn, we won't get our host-specific
files.

Note that we were never able to rcup/rcdn on a non-Andrew AFS machine.

With the symlinks, we can safely rcup/rcdn on Andrew AFS machines again.