shell: Fix some MANPATH problems
What I was noticing was that looking at the documentation for

    man git-rev-list

was taking me to the 2.24 version of `git` instead of the `2.44` version
of git. The 2.24 version was isntalled in `/usr/bin/git`, while the 2.44
version was installed in /usr/local/bin/git via Homebrew.

It also looks like the MANPATH environment variable was getting set to
read from `/usr/share` before `/usr/local/share`, because of what was in
`/etc/manpaths`. After I flipped those, I noticed that there were still
a bunch of other things in the MANPATH that seemed to only be because
`/etc/zprofile` runs `/usr/libexec/path_helper` for shells inside tmux,
but not those outside tmux (probably login vs non-login?).

In any case, it seems like without any of these, I'm able to find man
pages again so I'm guessing that something after Catalina fixed the
previous problem with man path searching, and that these explicit
MANPATH variable initializations are not required anymore.